+++ FmodForFoxes/FmodForFoxes/CoreSystem.cs
@@ -23,16 +23,34 @@ namespace FmodForFoxes
 		public static Sound LoadSound(string path)
 		{
 			var buffer = FileLoader.LoadFileAsBuffer(path);
+			return LoadSound(buffer);
+		}
 
+		/// <summary>
+		/// Loads sound from stream.
+		/// Use this function to load short sound effects.
+		/// </summary>
+		public static Sound LoadSound(Stream stream)
+		{
+			var buffer = FileLoader.LoadFileAsBuffer(stream);
+			return LoadSound(buffer);
+		}
+
+		/// <summary>
+		/// Loads sound from a buffer.
+		/// Use this function to load short sound effects.
+		/// </summary>
+		public static Sound LoadSound(byte[] buffer)
+		{
 			var info = new FMOD.CREATESOUNDEXINFO();
 			info.length = (uint)buffer.Length;
 			info.cbsize = Marshal.SizeOf(info);
 
 			Native.createSound(
-					buffer,
-					FMOD.MODE.OPENMEMORY | FMOD.MODE.CREATESAMPLE,
-					ref info,
-					out FMOD.Sound newSound
+				buffer,
+				FMOD.MODE.OPENMEMORY | FMOD.MODE.CREATESAMPLE,
+				ref info,
+				out FMOD.Sound newSound
 			);
 
 			return new Sound(newSound);
@@ -45,7 +63,25 @@ namespace FmodForFoxes
 		public static Sound LoadStreamedSound(string path)
 		{
 			var buffer = FileLoader.LoadFileAsBuffer(path);
+			return LoadStreamedSound(buffer);
+		}
 
+		/// <summary>
+		/// Loads streamed sound stream from file.
+		/// Use this function to load music and long ambience tracks.
+		/// </summary>
+		public static Sound LoadStreamedSound(Stream stream)
+		{
+			var buffer = FileLoader.LoadFileAsBuffer(stream);
+			return LoadStreamedSound(buffer);
+		}
+
+		/// <summary>
+		/// Loads streamed sound stream from file.
+		/// Use this function to load music and long ambience tracks.
+		/// </summary>
+		public static Sound LoadStreamedSound(byte[] buffer)
+		{
 			// Internal FMOD pointer points to this memory, so we don't want it to go anywhere.
 			var handle = GCHandle.Alloc(buffer, GCHandleType.Pinned);
 
@@ -54,10 +90,10 @@ namespace FmodForFoxes
 			info.cbsize = Marshal.SizeOf(info);
 
 			Native.createStream(
-					buffer,
-					FMOD.MODE.OPENMEMORY | FMOD.MODE.CREATESTREAM,
-					ref info,
-					out FMOD.Sound newSound
+				buffer,
+				FMOD.MODE.OPENMEMORY | FMOD.MODE.CREATESTREAM,
+				ref info,
+				out FMOD.Sound newSound
 			);
 
 			return new Sound(newSound, buffer, handle);